Use the correct version of emsdk for our fork
The emsdk I chose in #31 is wrong.

Apparently the emsdk build system had a big rewrite, and the new system
is from 1.38.33 on (unfortunately that's the version one I picked,
because it was the oldest one that I could easily find).

I read up on emscripten and realized that our old fork needed an even
older version - the one before the rewrite. And I figured out how to
install and activate those older versions: sdk-fastcomp-1.38.30-64bit

And while I'm at it, I blew away the emsdk directory. We shouldn't have
checked it in. I instead added a snippet to our build script to fetch it
from the official github repo as a zip if necessary.

When we bring our fork up to date with google's we will need to use a
newer version of emsdk, one that is after their build system rewrite.